Rename /genstream command to /profile-genstream and move to appropriate module
